+++ src/Reader/Json.php
@@ -10,6 +10,20 @@ namespace Laminas\Config\Reader;
 
 use Laminas\Config\Exception;
 
+use function array_replace_recursive;
+use function dirname;
+use function file_get_contents;
+use function is_array;
+use function is_file;
+use function is_readable;
+use function json_decode;
+use function json_last_error;
+use function json_last_error_msg;
+use function sprintf;
+use function trim;
+
+use const JSON_ERROR_NONE;
+
 /**
  * JSON config reader.
  */